+++ data_validation/clients.py
@@ -70,7 +70,7 @@ except Exception:
 try:
     from third_party.ibis.ibis_mssql import connect as mssql_connect
 except Exception:
-    mssql_connect = _raise_missing_client_error("pip install pymssql")
+    mssql_connect = _raise_missing_client_error("pip install pyodbc")
 
 try:
     from third_party.ibis.ibis_snowflake.client import (
@@ -152,15 +152,18 @@ def list_tables(client, schema_name):
         return client.list_tables()
 
 
-def get_all_tables(client):
+def get_all_tables(client, allowed_schemas=None):
     """Return a list of tuples with database and table names.
 
     client (IbisClient): Client to use for tables
+    allowed_schemas (List[str]): List of schemas to pull.
     """
     table_objs = []
     schemas = list_schemas(client)
 
     for schema_name in schemas:
+        if allowed_schemas and schema_name not in allowed_schemas:
+            continue
         try:
             tables = list_tables(client, schema_name)
         except Exception as e: