+++ data_validation/query_builder/query_builder.py
@@ -65,7 +65,13 @@ class AggregateField(object):
     @staticmethod
     def sum(field_name=None, alias=None):
         return AggregateField(
-            ibis.expr.api.NumericColumn.sum, field_name=field_name, alias=alias,
+            ibis.expr.api.IntegerColumn.sum, field_name=field_name, alias=alias,
+        )
+
+    @staticmethod
+    def bit_xor(field_name=None, alias=None):
+        return AggregateField(
+            ibis.expr.api.IntegerColumn.bit_xor, field_name=field_name, alias=alias,
         )
 
     def compile(self, ibis_table):
@@ -207,7 +213,8 @@ class ColumnReference(object):
 
 
 class CalculatedField(object):
-    def __init__(self, ibis_expr, config, fields):
+    def __init__(self, ibis_expr, config, fields, cast=None, **kwargs):
+
         """ A representation of an calculated field to build a query.
 
         Args:
@@ -217,17 +224,24 @@ class CalculatedField(object):
         self.expr = ibis_expr
         self.config = config
         self.fields = fields
-
-    @staticmethod
-    def hash(config, fields):
-        return CalculatedField(ibis.expr.api.ValueExpr.hash, config, fields,)
+        self.cast = cast
+        self.kwargs = kwargs
 
     @staticmethod
     def concat(config, fields):
         if config.get("default_concat_separator") is None:
             config["default_concat_separator"] = ibis.literal(",")
         fields = [config["default_concat_separator"], fields]
-        return CalculatedField(ibis.expr.api.StringValue.join, config, fields,)
+        cast = "string"
+        return CalculatedField(
+            ibis.expr.api.StringValue.join, config, fields, cast=cast,
+        )
+
+    @staticmethod
+    def hash(config, fields):
+        if config.get("default_hash_function") is None:
+            how = "farm_fingerprint"
+        return CalculatedField(ibis.expr.api.ValueExpr.hash, config, fields, how=how,)
 
     @staticmethod
     def ifnull(config, fields):
@@ -265,14 +279,16 @@ class CalculatedField(object):
             elif isinstance(field, list):
                 compiled_fields.append(self._compile_fields(ibis_table, field))
             else:
-                compiled_fields.append(ibis_table[field])
+                if self.cast:
+                    compiled_fields.append(ibis_table[field].cast(self.cast))
+                else:
+                    compiled_fields.append(ibis_table[field])
 
         return compiled_fields
 
     def compile(self, ibis_table):
         compiled_fields = self._compile_fields(ibis_table, self.fields)
-
-        calc_field = self.expr(*compiled_fields)
+        calc_field = self.expr(*compiled_fields, **self.kwargs)
         if self.config["field_alias"]:
             calc_field = calc_field.name(self.config["field_alias"])