+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Untagged
 
+## 1.1.8
+
+- Adding and documenting `find-tables` CLI feature with schema filter
+- Correct filter errors caused by SQL Alchemy errors
+- Adding beta calculated fields logic
+
 ## 1.1.7
 
 - Adding tests to validate BIGNUMERIC BQ type behavior